requirements.txt
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
python-dotenv>=1.0.0
boto3>=1.35.0
pytest>=8.0.0
